# Normalize deserialization errors for bad data